# API-WEB-SAUDE/src/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "API de informações de Clínicas e Hospitais",
		"description": "Sobre a API de Clínicas e Hospitais",
		"contact": {
			"email": "ericspsantos@gmail.com"
		},
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "http://localhost:5000",
			"description": "API WEB SAÚDE local"
		},
		{
			"url": "https://api-web-saude.vercel.app/",
			"description": "API WEB SAÚDE hospedada"
		}
	],
	"paths": {
...
		"/buscar": {
			"get": {
				"summary": "Buscar Consultas por Nome",
				"description": "Esta rota busca consultas médicas por nome.",
				"tags": ["Consultas"],
				"parameters": [
					{
						"name": "nome",
						"in": "query",
						"description": "Nome da consulta a ser buscada.",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Consultas encontradas com sucesso.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Consultas"
								}
							}
						}
					},
					"400": {
						"description": "Campos inválidos ou nome muito curto.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										},
										"Errors": {
											"type": "array",
											"items": {
												"type": "string"
											}
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/unidades-de-saude": {
			"get": {
				"summary": "Listar Unidades de Saúde",
				"description": "Esta rota lista unidades de saúde.",
				"tags": ["Consultas"],
				"responses": {
					"200": {
						"description": "Unidades de saúde listadas com sucesso.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Consultas"
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/especialidades/{nome}": {
			"get": {
				"summary": "Buscar Especialidades por Nome da Unidade de Saúde",
				"description": "Esta rota busca especialidades médicas pelo nome da unidade de saúde.",
				"tags": ["Consultas"],
				"parameters": [
					{
						"name": "nome",
						"in": "path",
						"description": "Nome da unidade de saúde para buscar especialidades.",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Especialidades encontradas com sucesso.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Consultas"
								}
							}
						}
					},
					"404": {
						"description": "Nenhuma especialidade encontrada para o nome fornecido.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/unidade-de-saude/{nome}": {
			"get": {
				"summary": "Buscar Unidade de Saúde por Nome",
				"description": "Esta rota busca uma unidade de saúde por nome.",
				"tags": ["Consultas"],
				"parameters": [
					{
						"name": "nome",
						"in": "path",
						"description": "Nome da unidade de saúde a ser buscada.",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Unidade de saúde encontrada com sucesso.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Consultas"
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/total-unidades-de-saude": {
			"get": {
				"summary": "Obter Total de Unidades de Saúde",
				"description": "Esta rota retorna o total de unidades de saúde.",
				"tags": ["Consultas"],
				"responses": {
					"200": {
						"description": "Total de unidades de saúde obtido com sucesso.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"total": {
											"type": "integer"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/pedidos": {
			"get": {
				"summary": "Listar Unidades de Saúde Pendentes",
				"description": "Esta rota lista unidades de saúde com pedidos pendentes.",
				"tags": ["Consultas"],
				"responses": {
					"200": {
						"description": "Unidades de saúde com pedidos pendentes listadas com sucesso.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
...
			"Consultas": {
				"type": "object",
				"description": "Este caminho contém operações para gerenciar consultas médicas relacionadas tanto a Clínicas como Hospitais.",
				"properties": {
					"id": {
						"type": "integer"
					},
					"nome": {
						"type": "string"
					},
					"descricao": {
						"type": "string"
					}
				}
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}